+++ battle-service/src/main/java/by/pashkevich/mikhail/service/SettingService.java
@@ -5,5 +5,5 @@ public interface SettingService {
 
     String getFixedRate();
 
-    Integer getMoveWaitingTime();
+    Long getMoveWaitingTime();
 }
